requirements.txt
openai
tiktoken 
chromadb
langchain
langchain-community
langchain-openai
flask
GitPython
python-dotenv
Flask
